gst-plugin-hexagon-nn: correct variable's datatype
variable kTimeOut has value larger than what can fit in the
current datatype. Hence changed it from uint8_t to uint32_t

+++ gst-plugin-hexagon-nn/engines/nnengine/nnengine.h
@@ -197,7 +197,7 @@ protected:
   NNDriver                 nn_driver_;
   void *                   outputs_[kMaxOut];
 
-  static const uint8_t     kTimeOut = 1000;
+  static const uint32_t     kTimeOut = 1000;
 
   std::future<void>        future_;
   std::mutex               process_lock_;
